Support 64-bit enum values and add bigenum example
Fix parser and generation to handle enum values wider than 32 bits: change State::enum_val to uint64_t and parse enum values with std::stoull. Stop filtering out values based on the declared underlying type size so large enum constants are emitted. Add a new examples/bigenum.cpp that reproduces the large-enum parsing case and wire it into CMake (CMakeLists.txt and cmake.toml), linking the example against sdkgenny and taocpp::pegtl. This enables testing and reproducing the previous "stoul argument out of range" behavior and ensures correct handling of 64-bit enum constants.

+++ include/sdkgenny_parser.hpp
@@ -163,7 +163,7 @@ struct State {
     std::string enum_type{};
     bool enum_class{};
     std::string enum_val_name{};
-    uint32_t enum_val{};
+    uint64_t enum_val{};
     std::vector<std::tuple<std::string, uint64_t>> enum_vals{};
 
     std::string struct_name{};
@@ -438,7 +438,7 @@ template <> struct Action<EnumExpr> {
 
 template <> struct Action<EnumVal> {
     template <typename Input> static void apply(const Input& in, State& s) {
-        s.enum_val = std::stoul(in.string(), nullptr, 0);
+        s.enum_val = std::stoull(in.string(), nullptr, 0);
     }
 };
